// src/api/generated/models/createUserRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Arco Enterprise Admin API
 * Spring Boot 管理后台示例契约，建议由 springdoc-openapi 的 /v3/api-docs 替换。
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserRequest {
  /**
     * @minLength 2
     * @maxLength 50
     */
  name: string;
  email: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  department: string;
  /** @minItems 1 */
  roles: string[];
}